fixed int atoi to long atoi
renamed tings

+++ libft/standard/ft_atoi.c
@@ -12,9 +12,9 @@
 
 #include "../includes/libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	res;
+	long	res;
 	int	minus;
 	int	i;
